Rename back to DLCStatus
